ref: Update t_philo struct to include *config member
The recent code changes involve updating the t_philo struct to include a *config member. This change allows for easier access to the configuration settings for each philosopher. The *config member will be used in various functions throughout the codebase to access the time_to_die value for monitoring and checking if a philosopher is dead.

+++ inc/philo.h
@@ -6,7 +6,7 @@
 /*   By: airyago <airyago@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/22 13:25:43 by airyago           #+#    #+#             */
-/*   Updated: 2024/08/22 19:40:16 by airyago          ###   ########.fr       */
+/*   Updated: 2024/08/23 11:31:08 by airyago          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ typedef struct s_config
 	size_t			time_to_die;
 	size_t			time_to_eat;
 	size_t			time_to_sleep;
-	int				num_times_to_eat;
+	size_t			num_times_to_eat;
 }	t_config;
 
 typedef struct s_philo
